requirements.txt
aiohttp==3.8.6
celery==5.3.4
django-admin-interface==0.26.1
django-cors-headers==4.3.0
django-stubs-ext==4.2.5
django-types==0.19.1
djangorestframework==3.14.0
docopt==0.6.2
gql==3.4.1
pip-chill==1.0.3
psycopg==3.1.12
psycopg-binary==3.1.12
redis==4.6.0
yarg==0.1.9
# aiosignal==1.3.1 # Installed as dependency for aiohttp
# amqp==5.1.1 # Installed as dependency for kombu
# asgiref==3.7.2 # Installed as dependency for django
# async-timeout==4.0.3 # Installed as dependency for aiohttp
# attrs==23.1.0 # Installed as dependency for aiohttp
# backoff==2.2.1 # Installed as dependency for gql
# billiard==4.1.0 # Installed as dependency for celery
# certifi==2023.7.22 # Installed as dependency for requests
# charset-normalizer==3.3.1 # Installed as dependency for aiohttp, requests
# click==8.1.7 # Installed as dependency for celery, click-didyoumean, click-plugins, click-repl
# click-didyoumean==0.3.0 # Installed as dependency for celery
# click-plugins==1.1.1 # Installed as dependency for celery
# click-repl==0.3.0 # Installed as dependency for celery
# django==4.2.6 # Installed as dependency for django-cors-headers, django-stubs-ext, djangorestframework
# django-colorfield==0.10.1 # Installed as dependency for django-admin-interface
# frozenlist==1.4.0 # Installed as dependency for aiohttp, aiosignal
# graphql-core==3.2.3 # Installed as dependency for gql
# idna==3.4 # Installed as dependency for requests, yarl
# kombu==5.3.2 # Installed as dependency for celery
# multidict==6.0.4 # Installed as dependency for aiohttp, yarl
# pillow==10.1.0 # Installed as dependency for django-colorfield
# prompt-toolkit==3.0.39 # Installed as dependency for click-repl
# python-dateutil==2.8.2 # Installed as dependency for celery
# python-slugify==8.0.1 # Installed as dependency for django-admin-interface
# pytz==2023.3.post1 # Installed as dependency for djangorestframework
# requests==2.31.0 # Installed as dependency for yarg
# six==1.16.0 # Installed as dependency for python-dateutil
# sqlparse==0.4.4 # Installed as dependency for django
# text-unidecode==1.3 # Installed as dependency for python-slugify
# types-psycopg2==2.9.21.15 # Installed as dependency for django-types
# typing-extensions==4.8.0 # Installed as dependency for django-stubs-ext, psycopg
# tzdata==2023.3 # Installed as dependency for celery
# urllib3==2.0.7 # Installed as dependency for requests
# vine==5.0.0 # Installed as dependency for amqp, celery, kombu
# wcwidth==0.2.8 # Installed as dependency for prompt-toolkit
# yarl==1.9.2 # Installed as dependency for aiohttp, gql
